test: move builder tests to its own file
