+++ src/wrapper/wrapper_builder.rs
@@ -127,28 +127,3 @@ impl Default for WebRegWrapperBuilder {
         Self::new()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use reqwest::Client;
-
-    use crate::wrapper::wrapper_builder::WebRegWrapperBuilder;
-
-    #[test]
-    fn fail_construct_wrapper() {
-        let wrapper = WebRegWrapperBuilder::new()
-            .with_client(Client::new())
-            .with_default_term("FA23")
-            .try_build_wrapper();
-        assert!(wrapper.is_none());
-    }
-
-    #[test]
-    fn success_construct_wrapper() {
-        let wrapper = WebRegWrapperBuilder::new()
-            .with_cookies("abc")
-            .with_default_term("FA23")
-            .try_build_wrapper();
-        assert!(wrapper.is_some());
-    }
-}
